+++ fearless-utils/src/main/java/jp/co/soramitsu/fearless_utils/encrypt/junction/Junction.kt
@@ -1,0 +1,3 @@
+package jp.co.soramitsu.fearless_utils.encrypt.junction
+
+data class Junction(val type: JunctionType, val chaincode: ByteArray)
