+++ fearless-utils/src/main/java/jp/co/soramitsu/fearless_utils/encrypt/junction/JunctionType.kt
@@ -1,0 +1,6 @@
+package jp.co.soramitsu.fearless_utils.encrypt.junction
+
+enum class JunctionType {
+    SOFT,
+    HARD
+}
