+++ fearless-utils/src/main/java/jp/co/soramitsu/fearless_utils/encrypt/keypair/ethereum/Bip32KeypairFactory.kt
@@ -6,8 +6,8 @@ import jp.co.soramitsu.fearless_utils.encrypt.keypair.KeypairFactory
 import jp.co.soramitsu.fearless_utils.encrypt.keypair.derivePublicKey
 import jp.co.soramitsu.fearless_utils.extensions.fromUnsignedBytes
 import jp.co.soramitsu.fearless_utils.extensions.requireOrException
-import jp.co.soramitsu.fearless_utils.junction.Junction
-import jp.co.soramitsu.fearless_utils.junction.JunctionType
+import jp.co.soramitsu.fearless_utils.encrypt.junction.Junction
+import jp.co.soramitsu.fearless_utils.encrypt.junction.JunctionType
 import jp.co.soramitsu.fearless_utils.scale.utils.toUnsignedBytes
 import java.math.BigInteger
