+++ fearless-utils/src/main/java/jp/co/soramitsu/fearless_utils/encrypt/keypair/ethereum/EthereumKeypairFactory.kt
@@ -2,7 +2,7 @@ package jp.co.soramitsu.fearless_utils.encrypt.keypair.ethereum
 
 import jp.co.soramitsu.fearless_utils.encrypt.keypair.Keypair
 import jp.co.soramitsu.fearless_utils.encrypt.keypair.generate
-import jp.co.soramitsu.fearless_utils.junction.Junction
+import jp.co.soramitsu.fearless_utils.encrypt.junction.Junction
 
 object EthereumKeypairFactory {
