+++ fearless-utils/src/main/java/jp/co/soramitsu/fearless_utils/encrypt/keypair/substrate/Common.kt
@@ -3,8 +3,8 @@ package jp.co.soramitsu.fearless_utils.encrypt.keypair.substrate
 import jp.co.soramitsu.fearless_utils.encrypt.keypair.Keypair
 import jp.co.soramitsu.fearless_utils.encrypt.keypair.KeypairFactory
 import jp.co.soramitsu.fearless_utils.hash.Hasher.blake2b256
-import jp.co.soramitsu.fearless_utils.junction.Junction
-import jp.co.soramitsu.fearless_utils.junction.JunctionType
+import jp.co.soramitsu.fearless_utils.encrypt.junction.Junction
+import jp.co.soramitsu.fearless_utils.encrypt.junction.JunctionType
 import jp.co.soramitsu.fearless_utils.scale.dataType.string
 import jp.co.soramitsu.fearless_utils.scale.dataType.toByteArray
