+++ fearless-utils/src/main/java/jp/co/soramitsu/fearless_utils/encrypt/keypair/substrate/Sr25519SubstrateKeypairFactory.kt
@@ -3,8 +3,8 @@ package jp.co.soramitsu.fearless_utils.encrypt.keypair.substrate
 import jp.co.soramitsu.fearless_utils.encrypt.Sr25519
 import jp.co.soramitsu.fearless_utils.encrypt.keypair.Keypair
 import jp.co.soramitsu.fearless_utils.encrypt.keypair.KeypairFactory
-import jp.co.soramitsu.fearless_utils.junction.Junction
-import jp.co.soramitsu.fearless_utils.junction.JunctionType
+import jp.co.soramitsu.fearless_utils.encrypt.junction.Junction
+import jp.co.soramitsu.fearless_utils.encrypt.junction.JunctionType
 
 class Sr25519Keypair(
     override val privateKey: ByteArray,
