+++ fearless-utils/src/main/java/jp/co/soramitsu/fearless_utils/encrypt/mnemonic/Mnemonic.kt
@@ -1,0 +1,19 @@
+package jp.co.soramitsu.fearless_utils.encrypt.mnemonic
+
+class Mnemonic(
+
+    val words: String,
+
+    val wordList: List<String>,
+
+    val entropy: ByteArray
+) {
+
+    enum class Length(val byteLength: Int) {
+        TWELVE(16),
+        FIFTEEN(20),
+        EIGHTEEN(24),
+        TWENTY_ONE(28),
+        TWENTY_FOUR(32);
+    }
+}
