+++ fearless-utils/src/main/java/jp/co/soramitsu/fearless_utils/encrypt/mnemonic/SeedFactory.kt
@@ -1,0 +1,10 @@
+package jp.co.soramitsu.fearless_utils.encrypt.mnemonic
+
+interface SeedFactory {
+
+    class Result(val seed: ByteArray, val mnemonic: Mnemonic)
+
+    fun createSeed(length: Mnemonic.Length, password: String?): Result
+
+    fun deriveSeed(mnemonicWords: String, password: String?): Result
+}
