+++ fearless-utils/src/main/java/jp/co/soramitsu/fearless_utils/encrypt/seed/ethereum/EthereumSeedFactory.kt
@@ -1,0 +1,23 @@
+package jp.co.soramitsu.fearless_utils.encrypt.seed.ethereum
+
+import jp.co.soramitsu.fearless_utils.encrypt.mnemonic.Mnemonic
+import jp.co.soramitsu.fearless_utils.encrypt.mnemonic.MnemonicCreator
+import jp.co.soramitsu.fearless_utils.encrypt.mnemonic.SeedCreator
+import jp.co.soramitsu.fearless_utils.encrypt.mnemonic.SeedFactory
+
+object EthereumSeedFactory : SeedFactory {
+
+    override fun createSeed(length: Mnemonic.Length, password: String?): SeedFactory.Result {
+        val mnemonic = MnemonicCreator.randomMnemonic(length)
+        val seed = SeedCreator.deriveSeed(mnemonic.words.encodeToByteArray(), password)
+
+        return SeedFactory.Result(seed, mnemonic)
+    }
+
+    override fun deriveSeed(mnemonicWords: String, password: String?): SeedFactory.Result {
+        val mnemonic = MnemonicCreator.fromWords(mnemonicWords)
+        val seed = SeedCreator.deriveSeed(mnemonic.words.encodeToByteArray(), password)
+
+        return SeedFactory.Result(seed, mnemonic)
+    }
+}
