--- fearless-utils/src/main/java/jp/co/soramitsu/fearless_utils/junction/Junction.kt
@@ -1,3 +1,0 @@
-package jp.co.soramitsu.fearless_utils.junction
-
-data class Junction(val type: JunctionType, val chaincode: ByteArray)
