--- fearless-utils/src/main/java/jp/co/soramitsu/fearless_utils/junction/JunctionType.kt
@@ -1,6 +1,0 @@
-package jp.co.soramitsu.fearless_utils.junction
-
-enum class JunctionType {
-    SOFT,
-    HARD
-}
