+++ fearless-utils/src/test/java/jp/co/soramitsu/fearless_utils/encrypt/keypair/ethereum/EthereumKeypairFactoryTest.kt
@@ -5,7 +5,7 @@ import jp.co.soramitsu.fearless_utils.encrypt.keypair.SeedTestCase
 import jp.co.soramitsu.fearless_utils.extensions.fromHex
 import jp.co.soramitsu.fearless_utils.extensions.toHexString
 import jp.co.soramitsu.fearless_utils.getResourceReader
-import jp.co.soramitsu.fearless_utils.junction.BIP32JunctionDecoder
+import jp.co.soramitsu.fearless_utils.encrypt.junction.BIP32JunctionDecoder
 import org.junit.Assert.*
 import org.junit.Test
