+++ fearless-utils/src/testShared/java/jp/co/soramitsu/fearless_utils/encrypt/mnemonic/MnemonicTestCase.kt
@@ -1,0 +1,10 @@
+package jp.co.soramitsu.fearless_utils.encrypt.mnemonic
+
+import com.google.gson.annotations.SerializedName
+
+data class MnemonicTestCase(
+    val mnemonic: String,
+    val path: String,
+    @SerializedName("pk")
+    val expectedPublicKey: String,
+)
